Restructuring part 2: Legacy and MPcules (now just "molecules") on API side (still need to update core and builders
