+++ emmet-api/emmet/api/routes/legacy/jcesr/resources.py
@@ -2,7 +2,7 @@ from maggma.api.resource import ReadOnlyResource
 from emmet.core.molecules_jcesr import MoleculesDoc
 
 from maggma.api.query_operator import PaginationQuery, SortQuery, SparseFieldsQuery
-from emmet.api.routes.molecules.query_operators import (
+from emmet.api.routes.legacy.jcesr.query_operators import (
     MoleculeBaseQuery,
     MoleculeElementsQuery,
     MoleculeFormulaQuery,
@@ -12,7 +12,7 @@ from emmet.api.core.global_header import GlobalHeaderProcessor
 from emmet.api.core.settings import MAPISettings
 
 
-def molecules_resource(molecules_store):
+def jcesr_resource(molecules_store):
     resource = ReadOnlyResource(
         molecules_store,
         MoleculesDoc,
@@ -26,7 +26,8 @@ def molecules_resource(molecules_store):
             SparseFieldsQuery(MoleculesDoc, default_fields=["task_id"]),
         ],
         header_processor=GlobalHeaderProcessor(),
-        tags=["Molecules"],
+        tags=["JCESR Electrolyte Genome"],
+        sub_path="/jcesr/",
         disable_validation=True,
         timeout=MAPISettings().TIMEOUT
     )
