+++ emmet-api/emmet/api/routes/mpcules/redox/resources.py
@@ -43,7 +43,7 @@ def redox_resource(redox_store):
             SparseFieldsQuery(RedoxDoc, default_fields=["molecule_id", "property_id", "solvent", "last_updated"],),
         ],
         header_processor=GlobalHeaderProcessor(),
-        tags=["MPcules Redox"],
+        tags=["Molecules Redox"],
         sub_path="/redox/",
         disable_validation=True,
         timeout=MAPISettings().TIMEOUT,
