+++ emmet-api/emmet/api/routes/mpcules/summary/resources.py
@@ -51,7 +51,7 @@ def summary_resource(summary_store):
         ],
         hint_scheme=SummaryHintScheme(),
         header_processor=GlobalHeaderProcessor(),
-        tags=["MPcules Summary"],
+        tags=["Molecules Summary"],
         disable_validation=True,
         timeout=timeout
     )
