+++ emmet-api/emmet/api/routes/mpcules/tasks/resources.py
@@ -40,7 +40,7 @@ def task_resource(task_store):
         ],
         header_processor=GlobalHeaderProcessor(),
         hint_scheme=TasksHintScheme(),
-        tags=["MPcules Tasks"],
+        tags=["Molecules Tasks"],
         sub_path="/tasks/",
         timeout=timeout,
         disable_validation=True,
@@ -54,7 +54,7 @@ def task_deprecation_resource(task_store):
         task_store,
         DeprecationDoc,
         query_operators=[DeprecationQuery(), PaginationQuery()],
-        tags=["MPcules Tasks"],
+        tags=["Molecules Tasks"],
         enable_get_by_key=False,
         enable_default_search=True,
         sub_path="/tasks/deprecation/",
